backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
sentence-transformers
numpy
scikit-learn
easyocr
python-multipart
networkx
pytest
httpx
pytest-asyncio
google-generativeai

